change timeout to double

+++ ssc_interface_wrapper_ros2/include/ssc_interface_wrapper/ssc_interface_wrapper_config.hpp
@@ -27,7 +27,7 @@ namespace ssc_interface_wrapper
   struct Config
   {
     // Timeout for command receipt in seconds
-    int controller_timeout = 1;
+    double controller_timeout = 1;
 
     // Stream operator for this config
     friend std::ostream &operator<<(std::ostream &output, const Config &c)
